Убрал трекинг dist/js/
