--- Flora/index.html
@@ -1,149 +1,0 @@
-<!DOCTYPE html>
-<html lang="ru">
-<head>
-    <title>FloraTruck</title>
-    <meta charset="utf-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1" />
-    <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/normalize/8.0.1/normalize.min.css">
-    <link rel="stylesheet" href="css/style.css">
-    <link rel="preconnect" href="https://fonts.googleapis.com">
-    <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
-    <script src="https://api-maps.yandex.ru/2.1/?lang=ru_RU&amp;apikey=<ваш API-ключ>" type="text/javascript"></script>
-    <script src="js/custom_search.js" type="text/javascript"></script>
-    <link href="https://fonts.googleapis.com/css2?family=Inter:wght@700&display=swap" rel="stylesheet">
-</head>
-<body>
-<section class="main">
-    <header>
-        <div class="header">
-            <div class="logo">
-                <img src="image/logo.png" alt="FloraTrack">
-            </div>
-            <nav>
-                <ul class="nav">
-                    <li class="item"><a class="link"    href="#">О компании</a></li>
-                    <li class="item"><a class="link"    href="">Новости компании</a></li>
-                    <li class="item"><a class="link"    href="#">Контакты</a></li>
-                </ul>
-            </nav>
-<div class="lk">
-    <h2 class="fran"><span class="vec"><img src="image/Vector.png" alt=""></span><a class="f_link" href="">Франшиза</a></h2>
-    <h2 class="fran"><span class="vec"><img src="image/Vector1.png" alt=""></span><a class="f_link" href="">Личный кабинет</a></h2>
-
-</div>
-
-
-
-
-        </div>
-    </header>
-    <div class="box_about">
-        <div class="box_item">
-            <img class="box_top" src="image/kargo.png" alt="Авиа">
-            <h2 class="text_it"><a class="f_link_r" href="">Авиаперевозки карго <span class="right"><img class="box_front" src="image/arrow.png" alt=""></span></a></h2>
-
-        </div>
-
-            <div class="box_item">
-            <img class="box_top" src="image/logist.png" alt="Авиа">
-                <h2 class="text_it"><a class="f_link_r" href="">Транспортная логистика <span class="right"><img class="box_front" src="image/arrow.png" alt=""></span></a></h2>
-            </div>
-    </div>
-    <div class="btn_scroll">
-        <img src="image/mouse.png" alt="Scroll">
-        <h2 class="ns">Больше о нас</h2>
-    </div>
-    <div class="social-icon">
-       <div class="vr">
-        <div class="v">
-            <a class="fran1" href="https://chat.whatsapp.com/Ir7tgRBshX11cfP7aRskr0"> Чат WhatsApp
-          <img src="image/chat.png"></a>
-       </div>
-        <div class="v1">
-
-            <a href=""><img src="image/chat2.png"></a>
-        </div> </div>
-    </div>
-</section>
-<section class="flora">
-<div class="about">
-<p class="pere">Flora truck - это Крупнейший авиа экспедитор и один из самых крупных перевозчиков цветочной продукции в России</p>
-    <p class="cvet">Доставка цветочной продукции с любых аукционов и плантаций: Эквадора, Колумбии, Голландии, Кении, Израиля, Коста-Рики</p>
-    <p class="struk">Мы имеем развитую инфраструктуру логистических комплексов, сеть которых постоянно расширяется, благодаря этому мы можем обеспечить доставку товара в любую точку России .</p>
-</div>
-</section>
-<section class="lider">
-
-
-
-    <div class="ld"> <div class="wrapper">
-        <p class="truck">FloraTruck - компания лидер и надежный партнер</p>
-        <div class="services">
-            <div class="services_item">
-                <h3 class="title_h3"> >500 т</h3>
-                <p class="services_text">собственных консолидированных вылетов  ежемесячно</p>
-            </div>
-
-            <div class="services_item">
-                <h3 class="title_h3"> >150 фур</h3>
-                <p class="services_text"> цветочной продукции обрабатываются нашими складами каждый месяц</p>
-            </div>
-
-            <div class="services_item">
-                <h3 class="title_h3"> 24/7</h3>
-                <p class="services_text"> работа компании</p>
-            </div>
-
-            <div class="services_item">
-                <h3 class="title_h3"> 3</h3>
-                <p class="services_text"> склада на территории России, сеть которых постоянно растет</p>
-            </div>
-
-
-        </div>
-
-    </div>
-    </div>
-</section>
-<section class="instagram">
-    <div class="wrapper">
-<p class="inst">Последние новости и события в нашем инстаграмм <a href="https://www.instagram.com/floratruck/"><span class="f">@FLORA</span><span class="t">TRACK</span></a></p>
-    </div>
-</section>
-<section class="sklad">
-    <div id="map">
-        <h3 class="vb">Выберите удобный склад</h3>
-    </div>
-</section>
-<section class="footer">
-    <div class="wrapperf">
-        <div class="kontak">
-            <div class="raspol">
-                <p class="cl">Главный офис</p>
-                <p class="rz">МО, г. Котельники</p>
-                   <p class="ad">Угрешский Проезд, с3к</p>
-
-                <a class="pereh" href="">Смотреть на карте</a>
-                <div class="tel">
-                    <a class="tt" href="tel: +74951234567">+7 (495) 123-45-67</a>
-                    <a class="te" href="mailto: example@mail.ru">example@mail.ru</a>
-                </div>
-            </div>
-            <div class="uslug">
-                <p class="cl">Услуги</p>
-                <a class="pereh" href="pereh">Транспортная логистика</a>
-                <a class="pereh" href="">авиаперевозки карго</a>
-                <h2 class="fran2"><a class="f_link1" href="">Личный кабинет</a></h2>
-            </div>
-            <h2 class="fran"><span class="vec"><img src="image/Vector.png" alt=""></span><a class="f_link" href="">Скачать</a></h2>
-
-        </div>
-        <div class="w">
-            <h3 class="obs"><a class="vopr" href="">Есть вопросы? Давайте обсудим <span><img src="image/pere.png" alt=""></span></a></h3>
-            <h3 class=""><a class="vopr" href="">Подпишись <span><img src="image/inst.png" alt=""></span></a></h3>
-
-        </div>
-    </div>
-</section>
-</body>
-</html>
