+++ webpack.config.js
@@ -1,0 +1,9 @@
+const path = require('path')
+
+module.exports = {
+    entry: './src/index.js',
+    output: {
+        path: path.resolve(__dirname, 'dist/js'),
+        filename: 'bundle.js'
+    }
+}
